+++ plugins/modules/keycloak_clientsecret_info.py
@@ -115,7 +115,8 @@ clientsecret_info:
 '''
 
 from ansible_collections.community.general.plugins.module_utils.identity.keycloak.keycloak import (
-    KeycloakAPI, KeycloakError, camel, get_token, keycloak_argument_spec,
+    KeycloakAPI, KeycloakError, get_token)
+from ansible_collections.community.general.plugins.module_utils.identity.keycloak.keycloak_clientsecret import (
     keycloak_clientsecret_module, keycloak_clientsecret_module_resolve_params)
